requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
PyGithub
pydantic-settings
